Extract actions from status buffer so we can call them from the help popup
